pylint: Ignore 'unknown option' on older pylint versions
As we might have to deal with different pylint versions for current and
upcomming Ansible tests, some options may be unknown on some pylint
versions used.

By disabling warning W0012, we are able to use different versions of
pylint and prepare the code for future checks while not breaking current
ones.