Change hashbang of Python scripts to python3
As described in PEP 394 there can be confusion when a script is executed
with the command "python". Many of the scripts changed in this commit are
Python3 (e.g. they call subprocess.run). On some systems the python
command is simply unavailable. On most distributions it is better to use
the command "python3".