module_utils: Add support for hostname datatype
One of the issues on ensuring idempotence behavior is that FreeIPA does
not expose datatypes along with the values stored in objects.

By treating data returned from FreeIPA API with the proper datatype, it
is easier to ensure that comparisons that will trigger changes (or not)
are properly executed, despite of the stored value or the provided
parameter data.

This patch provides an initial implementation of the datatypes used by
ansible-freeipa modules. An implementation for 'hostname' which ensures
data provided for hostnames contain a fully qualified hostname, always
in lower case, and an implementation of a 'list_of' to allow for
parameters which are lists of objects of a specific datatype.