ansbile_module_utils: Use IPA object data type in compare_args_ipa
When comparing a module argument set against an IPA argument set, for
each attribute, if it is a list or tuple, a set is created and the sets
are compared, this is more efficient than comparing the objects
individually, as the list order may be different, and using a set,
ordering is ignored. If the parameter value is a scalar, they are
directly compared.

Although efficient, this method causes idempotence issues with some data
types, for example, when the comparison of two strings must be case
insensitive.

The solution proposed here is to avoid direct value comparison by always
using the set comparison method, and that the data type of the values
stored in the sets can perform the proper comparison for the specific
data type.

To ensure the comparison is done only with sets, scalar values are
"wrapped" in a list, and the list is later converted to a set.

To cope with specific data type comparison (for example: hostnames,
services, or case insensitive strings) a new parameter was added, so a
dictionaire mapping an argument to a conversion function can be passed,
and the set of values is created by applying the conversion function to
the IPA and the argument values.

The complexity order of the comparison algorithm is not changed, but
data conversion, and possibly a customized comparison, are new
operations performed, which may have some impact on the execution time.