ansible_module_utils: Make list generation function data type aware
The list generation functions for adding and deleting members from IPA
objects create sets but are not aware of the underlying attribute data
type. Some data types need to perform special comparison, like case
insensitive string comparison, or ensure the existence of a FQDN.

This patch adds a new optional attribute to the list generation
functions, 'attr_datatype', that takes a function able to convert the
expected data (usually a string) to an object that will correctly handle
hashing and data comparison.

The functions were also modified to perform the list generation, but
still return the same data provided to the caller, withot conversion,
minimizing the required amount of code changes.