depolyment roles: Allow deployment without gather_facts
When using the deployment roles, some facts about the target node
distribution are required, but using 'gather_facts: true' takes a long
time as it probes for a lot of facts which are not used by the
deployment roles.

By restricting the facts loaded to the ones that are used, there's a
reduction in the deployment tasks, as there's less data to be probed
before the role is executed.