ipaclient: Use hostvars data instead of inventory group name
When using 'ansible_host' to override the target node name in an Ansible
inventory, the ipaclient deployment role fails as it is not able to
either correctly resolve the server name, or fail to join the host to
the realm, since the inventory name may match the actual hostname.

By retrieving the 'ansible_host' from 'hostvars' a user is allowed to
change the actual hostname without affecting the Ansible node used. Such
a setup is useful if one have a template inventory, or to setup nodes
using containers.