ipaclient: Use hostvars data instead of inventory group name
When using 'ansible_host' to override the target node name in an Ansible
inventory, the deployment role fails as it is not able to resolve the
server name, since the inventory name does not match the actual
hostname.

By retrieving the 'ansible_host' from 'hostvars' a user is allowed to
change the actual hostname as a varible, being able to use the inventory
as a template.