upstream ci: Fix ansible-lint errors due to missing collections
When running ansible-lint it tries to load and syntax check the
playbooks it evaluates, and if any of the modules used in the playbooks
can't be found, it fails with a fatal error, as if there was a syntax
error.

By adding a 'requirements.yml' file with all the collections that are
used in any playbook (be it a runtime, test, or development dependency),
forces ansible-lint to install and use the collections in its virtual
environment, successfully parsing the playbook files.

See more at https://github.com/ansible/ansible-lint/discussions/4048 and
https://github.com/ansible/ansible-lint/pull/4049.