lint tools: Pin ansible-lint and bump version of other tools
Due to a release of ansible-lint version 24.y, which is currently
broken, ansible-lint version is pinned to the minimum accepted version
for Ansible Galaxy, which is version 6.22, on the upstream CI checks,
PIP requirements file, and pre-commit environment.

Other lint versions like Flake8, Pylint and yamllint were updated to
newer versions to cope with false positives, and preparation for future
development.