c10s: Prepare environment for beta testing CentOS Stream 10
Provides build scripts and environment setup for preparing container
images to test ansible-freeipa using CentOS Stream 1r containeres.

Currently, the c10s containers are in pre-alpha stage, so these tests
should be executed only on a case by case basis, and the image should
also be created when needed and not on the regular image build pipeline.