upstream ci: Use util scripts instead of molecule
This patch start the removal of the dependency on molecule for tests,
by using the scripts under `utils` to setup the environment.

By not using molecule, we have more flexibility on using either docker
or podman as the container engine, and makes it easy to reproduce the
environment on different distros, allowing for a more consistent error
reproduction off Azure.