TEMP: ipagroup: Fix capitalization idempotence issue with multiple objects
When setting multiple objects in the same tasks, some validation and
proper capitalization of parameter values were skipped.

By using the EntryFactory to extract the entries data both issues are
fixed.

New test added to ipagroup suite:

    tests/group/test_groups_member_case_insensitive.yml

Note: This is a temporary commit that should be part of another PR. It's
been added here to show the feasiability of the EntryFactory class.
Tests for case insensitive comparison must be added before this change
is merged.