Enable deployment roles testing using Github workflows
This patch enables testing deployment roles using FreeIPA-Cluster-Test
action (https://github.com/rjeffman/freeipa/cluster-tests).

The tests require a cluster configuration to be used with ipalab-config
(https://pypi.org/project/ipalab-config) and a list of Ansible
playbooks. To ensure the proper host runs the tests, 'ipa-test-host' is
set using a vars file, and passed to the action.

The tests run on Github actions, only if one of the deployment roles are
modified by the PR.